+++ cli/src/revset_util.rs
@@ -346,7 +346,7 @@ pub fn parse_remote_auto_track_bookmarks_map(
         let Some(text) = &settings.auto_track_bookmarks else {
             continue;
         };
-        let expr = parse_remote_auto_track_text(ui, name, text, "auto-track-bookmarks")?;
+        let expr = parse_remote_string_expression(ui, name, text, "auto-track-bookmarks")?;
         matchers.insert(name.clone(), expr.to_matcher());
     }
     Ok(matchers)
@@ -364,7 +364,7 @@ pub fn parse_remote_auto_track_bookmarks_map_for_new_bookmarks(
     for (name, settings) in remote_settings {
         let mut exprs = Vec::new();
         if let Some(text) = &settings.auto_track_bookmarks {
-            exprs.push(parse_remote_auto_track_text(
+            exprs.push(parse_remote_string_expression(
                 ui,
                 name,
                 text,
@@ -372,7 +372,7 @@ pub fn parse_remote_auto_track_bookmarks_map_for_new_bookmarks(
             )?);
         }
         if let Some(text) = &settings.auto_track_created_bookmarks {
-            exprs.push(parse_remote_auto_track_text(
+            exprs.push(parse_remote_string_expression(
                 ui,
                 name,
                 text,
@@ -390,7 +390,33 @@ pub fn parse_remote_auto_track_bookmarks_map_for_new_bookmarks(
     Ok(matchers)
 }
 
-fn parse_remote_auto_track_text(
+/// Parses the given `remotes.<name>.fetch-bookmarks` setting.
+pub fn parse_remote_fetch_bookmarks(
+    ui: &Ui,
+    remote_settings: &RemoteSettingsMap,
+    name: &RemoteName,
+) -> Result<Option<StringExpression>, CommandError> {
+    remote_settings
+        .get(name)
+        .and_then(|settings| settings.fetch_bookmarks.as_ref())
+        .map(|text| parse_remote_string_expression(ui, name, text, "fetch-bookmarks"))
+        .transpose()
+}
+
+/// Parses the given `remotes.<name>.fetch-tags` setting.
+pub fn parse_remote_fetch_tags(
+    ui: &Ui,
+    remote_settings: &RemoteSettingsMap,
+    name: &RemoteName,
+) -> Result<Option<StringExpression>, CommandError> {
+    remote_settings
+        .get(name)
+        .and_then(|settings| settings.fetch_tags.as_ref())
+        .map(|text| parse_remote_string_expression(ui, name, text, "fetch-tags"))
+        .transpose()
+}
+
+fn parse_remote_string_expression(
     ui: &Ui,
     name: &RemoteName,
     text: &str,