+++ cli/tests/test_git_fetch.rs
@@ -245,6 +245,68 @@ fn test_git_fetch_multiple_remotes() {
 }
 
 #[test]
+fn test_git_fetch_default_bookmarks_and_tags() {
+    let test_env = TestEnvironment::default();
+    test_env.run_jj_in(".", ["git", "init", "repo"]).success();
+    let work_dir = test_env.work_dir("repo");
+
+    let rem1_repo = add_git_remote(&test_env, &work_dir, "rem1");
+    let rem2_repo = add_git_remote(&test_env, &work_dir, "rem2");
+    let setup_opid = work_dir.current_operation_id();
+
+    git::add_commit(&rem1_repo, "refs/tags/tag1", "file", b"", "1a", &[]);
+    git::add_commit(&rem1_repo, "refs/heads/branch1", "file", b"", "1b", &[]);
+    git::add_commit(&rem1_repo, "refs/tags/tag2", "file", b"", "1c", &[]);
+    git::add_commit(&rem1_repo, "refs/heads/branch2", "file", b"", "1d", &[]);
+    git::add_commit(&rem2_repo, "refs/tags/tag1", "file", b"", "2a", &[]);
+    git::add_commit(&rem2_repo, "refs/heads/branch1", "file", b"", "2b", &[]);
+    git::add_commit(&rem2_repo, "refs/tags/tag2", "file", b"", "2c", &[]);
+    git::add_commit(&rem2_repo, "refs/heads/branch2", "file", b"", "2d", &[]);
+
+    // Per-remote default config
+    test_env.add_config(indoc! {"
+        [remotes.rem1]
+        fetch-bookmarks = 'branch1'
+        fetch-tags = 'tag1'
+        [remotes.rem2]
+        fetch-bookmarks = 'branch2'
+        fetch-tags = 'tag2'
+    "});
+    let output = work_dir.run_jj(["git", "fetch", "--all-remotes"]);
+    insta::assert_snapshot!(output, @"
+    ------- stderr -------
+    bookmark: branch1@rem1 [new] untracked
+    bookmark: branch2@rem2 [new] untracked
+    tag: tag1@rem1 [new] 
+    tag: tag2@rem2 [new] 
+    [EOF]
+    ");
+
+    // Default fetch-bookmarks/tags should be disabled by --branch
+    work_dir.run_jj(["op", "restore", &setup_opid]).success();
+    let output = work_dir.run_jj(["git", "fetch", "--all-remotes", "--branch=*"]);
+    insta::assert_snapshot!(output, @"
+    ------- stderr -------
+    bookmark: branch1@rem1 [new] untracked
+    bookmark: branch1@rem2 [new] untracked
+    bookmark: branch2@rem1 [new] untracked
+    bookmark: branch2@rem2 [new] untracked
+    bookmark: rem1@rem1    [new] untracked
+    bookmark: rem2@rem2    [new] untracked
+    [EOF]
+    ");
+
+    // Default fetch-bookmarks/tags should be disabled by --tag
+    work_dir.run_jj(["op", "restore", &setup_opid]).success();
+    let output = work_dir.run_jj(["git", "fetch", "--all-remotes", "--tag=~*"]);
+    insta::assert_snapshot!(output, @"
+    ------- stderr -------
+    Nothing changed.
+    [EOF]
+    ");
+}
+
+#[test]
 fn test_git_fetch_with_ignored_refspecs() {
     let test_env = TestEnvironment::default();
     test_env.run_jj_in(".", ["git", "init", "repo"]).success();